+++ Makefile
@@ -6,7 +6,7 @@
 #    By: mweverli <mweverli@student.codam.n>          +#+                      #
 #                                                    +#+                       #
 #    Created: 2022/10/01 17:54:19 by mweverli      #+#    #+#                  #
-#    Updated: 2023/03/01 14:51:21 by mverbrug      ########   odam.nl          #
+#    Updated: 2023/03/15 16:28:56 by mverbrug      ########   odam.nl          #
 #                                                                              #
 # **************************************************************************** #
 
@@ -27,19 +27,20 @@ SRC			:=	\
 				utils/minishell_error.c
 
 VER_SRC		:=	$(SRC) \
-				tmp_mares/mini_parse.c		\
-				tmp_mares/parse_utils.c		\
-				builtin/builtin.c			\
-				builtin/cd.c				\
-				builtin/echo.c				\
-				builtin/env.c				\
-				builtin/exit.c				\
-				builtin/export.c			\
-				builtin/pwd.c				\
-				builtin/unset.c				\
-				executor/env_var_list.c		\
-				executor/env_var.c			\
-				executor/init_shell.c		\
+				tmp_mares/mini_parse.c				\
+				tmp_mares/parse_utils.c				\
+				builtin/builtin.c					\
+				builtin/cd.c						\
+				builtin/echo.c						\
+				builtin/env.c						\
+				builtin/exit.c						\
+				builtin/export.c					\
+				builtin/pwd.c						\
+				builtin/unset.c						\
+				env_var_init_shell/env_var_list.c	\
+				env_var_init_shell/env_var.c		\
+				env_var_init_shell/init_shell.c		\
+				executor/executor.c					\
 				test/ver_main.c
 WEV_SRC		:=	$(SRC) \
 				tmp_martijn/print_utils.c	\
