+++ include/minishell.h
@@ -56,7 +56,7 @@ typedef struct s_cmd
 	char			*cmd;
 	char			**args;
 	int				amount_of_args;
-//	t_redir			*redir;
+	// t_redir			*redir;
 	struct s_cmd	*next;
 }				t_cmd;
