+++ src/env_var_init_shell/env_var_list.c
