+++ src/env_var_init_shell/init_shell.c
